```
- The first line is the commit message. This is a short summary of what is being committed.
- The next line is blank.
- The next line is the branch you are on and if it is up to date with the remote branch.
- The next line is blank.
- The next line is a list of changes that are being committed.

In this case, we are deleting some files and adding some new files. We also have some files that are being modified.

- The next line is blank.
- The next line is blank.
- The next line is blank.

We can now add our commit message. This message should be a short summary of what is being committed. It should be in the present tense. It should be written in the imperative mood. It should be less than 50 characters. It should not end with a period. It should be capitalized.